Update MRAA timer functions (#950)
* Update MRAA timer functions

- Found MRAA timer functions not working properly
- Copy timing functions from SPIDEV

+++ utility/MRAA/compatibility.h
@@ -15,6 +15,7 @@ extern "C" {
 #include <stddef.h>
 #include <time.h>
 #include <sys/time.h>
+#include <stdint.h>
 
 void __msleep(int milisec);
 
@@ -22,7 +23,7 @@ void __usleep(int milisec);
 
 void __start_timer();
 
-long __millis();
+uint32_t __millis();
 
 #ifdef __cplusplus
 }
